LU-4023 build: wrong type used
Fixed implicit conversion from 'unsigned long long' to 'int'.

+++ lustre/lov/lov_request.c
@@ -203,12 +203,12 @@ out:
 }
 
 extern void osc_update_enqueue(struct lustre_handle *lov_lockhp,
-                               struct lov_oinfo *loi, int flags,
-                               struct ost_lvb *lvb, __u32 mode, int rc);
+			       struct lov_oinfo *loi, __u64 flags,
+			       struct ost_lvb *lvb, __u32 mode, int rc);
 
 static int lov_update_enqueue_lov(struct obd_export *exp,
 				  struct lustre_handle *lov_lockhp,
-				  struct lov_oinfo *loi, int flags, int idx,
+				  struct lov_oinfo *loi, __u64 flags, int idx,
 				  struct ost_id *oi, int rc)
 {
 	struct lov_obd *lov = &exp->exp_obd->u.lov;
@@ -455,7 +455,7 @@ out_set:
         RETURN(rc);
 }
 
-int lov_fini_match_set(struct lov_request_set *set, __u32 mode, int flags)
+int lov_fini_match_set(struct lov_request_set *set, __u32 mode, __u64 flags)
 {
         int rc = 0;
         ENTRY;
